+++ packages/interface-compliance-tests/CHANGELOG.md
@@ -75,6 +75,26 @@
   * dependencies
     * @libp2p/utils bumped from ^5.2.4 to ^5.2.5
 
+## [6.4.4](https://github.com/libp2p/js-libp2p/compare/interface-compliance-tests-v6.4.3...interface-compliance-tests-v6.4.4) (2025-04-16)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/crypto bumped from ^5.1.0 to ^5.1.1
+    * @libp2p/echo bumped from ^2.1.19 to ^2.1.20
+    * @libp2p/interface bumped from ^2.8.0 to ^2.9.0
+    * @libp2p/interface-internal bumped from ^2.3.10 to ^2.3.11
+    * @libp2p/logger bumped from ^5.1.14 to ^5.1.15
+    * @libp2p/memory bumped from ^1.1.5 to ^1.1.6
+    * @libp2p/multistream-select bumped from ^6.0.21 to ^6.0.22
+    * @libp2p/peer-collections bumped from ^6.0.26 to ^6.0.27
+    * @libp2p/peer-id bumped from ^5.1.1 to ^5.1.2
+    * @libp2p/plaintext bumped from ^2.0.21 to ^2.0.22
+    * @libp2p/utils bumped from ^6.6.1 to ^6.6.2
+    * libp2p bumped from ^2.8.3 to ^2.8.4
+
 ## [6.4.3](https://github.com/libp2p/js-libp2p/compare/interface-compliance-tests-v6.4.2...interface-compliance-tests-v6.4.3) (2025-04-09)