+++ packages/metrics-devtools/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## [1.2.14](https://github.com/libp2p/js-libp2p/compare/devtools-metrics-v1.2.13...devtools-metrics-v1.2.14) (2025-04-16)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^2.8.0 to ^2.9.0
+    * @libp2p/interface-internal bumped from ^2.3.10 to ^2.3.11
+    * @libp2p/logger bumped from ^5.1.14 to ^5.1.15
+    * @libp2p/peer-id bumped from ^5.1.1 to ^5.1.2
+    * @libp2p/simple-metrics bumped from ^1.3.8 to ^1.3.9
+  * devDependencies
+    * @libp2p/crypto bumped from ^5.1.0 to ^5.1.1
+
 ## [1.2.13](https://github.com/libp2p/js-libp2p/compare/devtools-metrics-v1.2.12...devtools-metrics-v1.2.13) (2025-04-09)