+++ packages/peer-record/CHANGELOG.md
@@ -60,6 +60,18 @@
   * dependencies
     * @libp2p/utils bumped from ^5.2.4 to ^5.2.5
 
+## [8.0.27](https://github.com/libp2p/js-libp2p/compare/peer-record-v8.0.26...peer-record-v8.0.27) (2025-04-16)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/crypto bumped from ^5.1.0 to ^5.1.1
+    * @libp2p/interface bumped from ^2.8.0 to ^2.9.0
+    * @libp2p/peer-id bumped from ^5.1.1 to ^5.1.2
+    * @libp2p/utils bumped from ^6.6.1 to ^6.6.2
+
 ## [8.0.26](https://github.com/libp2p/js-libp2p/compare/peer-record-v8.0.25...peer-record-v8.0.26) (2025-04-09)