+++ packages/pubsub-floodsub/CHANGELOG.md
@@ -98,6 +98,22 @@
   * devDependencies
     * @libp2p/interface-compliance-tests bumped from ^5.3.0 to ^5.3.1
 
+## [10.1.33](https://github.com/libp2p/js-libp2p/compare/floodsub-v10.1.32...floodsub-v10.1.33) (2025-04-16)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^2.8.0 to ^2.9.0
+    * @libp2p/pubsub bumped from ^10.1.9 to ^10.1.10
+  * devDependencies
+    * @libp2p/crypto bumped from ^5.1.0 to ^5.1.1
+    * @libp2p/interface-compliance-tests bumped from ^6.4.3 to ^6.4.4
+    * @libp2p/logger bumped from ^5.1.14 to ^5.1.15
+    * @libp2p/peer-collections bumped from ^6.0.26 to ^6.0.27
+    * @libp2p/peer-id bumped from ^5.1.1 to ^5.1.2
+
 ## [10.1.32](https://github.com/libp2p/js-libp2p/compare/floodsub-v10.1.31...floodsub-v10.1.32) (2025-04-09)