+++ packages/stream-multiplexer-mplex/CHANGELOG.md
@@ -78,6 +78,19 @@
   * devDependencies
     * @libp2p/interface-compliance-tests bumped from ^5.3.0 to ^5.3.1
 
+## [11.0.34](https://github.com/libp2p/js-libp2p/compare/mplex-v11.0.33...mplex-v11.0.34) (2025-04-16)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^2.8.0 to ^2.9.0
+    * @libp2p/utils bumped from ^6.6.1 to ^6.6.2
+  * devDependencies
+    * @libp2p/interface-compliance-tests bumped from ^6.4.3 to ^6.4.4
+    * @libp2p/logger bumped from ^5.1.14 to ^5.1.15
+
 ## [11.0.33](https://github.com/libp2p/js-libp2p/compare/mplex-v11.0.32...mplex-v11.0.33) (2025-04-09)