+++ packages/transport-circuit-relay-v2/CHANGELOG.md
@@ -101,6 +101,24 @@
   * devDependencies
     * @libp2p/interface-compliance-tests bumped from ^5.3.0 to ^5.3.1
 
+## [3.2.10](https://github.com/libp2p/js-libp2p/compare/circuit-relay-v2-v3.2.9...circuit-relay-v2-v3.2.10) (2025-04-16)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/crypto bumped from ^5.1.0 to ^5.1.1
+    * @libp2p/interface bumped from ^2.8.0 to ^2.9.0
+    * @libp2p/interface-internal bumped from ^2.3.10 to ^2.3.11
+    * @libp2p/peer-collections bumped from ^6.0.26 to ^6.0.27
+    * @libp2p/peer-id bumped from ^5.1.1 to ^5.1.2
+    * @libp2p/peer-record bumped from ^8.0.26 to ^8.0.27
+    * @libp2p/utils bumped from ^6.6.1 to ^6.6.2
+  * devDependencies
+    * @libp2p/interface-compliance-tests bumped from ^6.4.3 to ^6.4.4
+    * @libp2p/logger bumped from ^5.1.14 to ^5.1.15
+
 ## [3.2.9](https://github.com/libp2p/js-libp2p/compare/circuit-relay-v2-v3.2.8...circuit-relay-v2-v3.2.9) (2025-04-09)