+++ packages/transport-tcp/CHANGELOG.md
@@ -89,6 +89,18 @@
   * devDependencies
     * @libp2p/interface-compliance-tests bumped from ^5.3.0 to ^5.3.1
 
+## [10.1.10](https://github.com/libp2p/js-libp2p/compare/tcp-v10.1.9...tcp-v10.1.10) (2025-04-16)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^2.8.0 to ^2.9.0
+    * @libp2p/utils bumped from ^6.6.1 to ^6.6.2
+  * devDependencies
+    * @libp2p/logger bumped from ^5.1.14 to ^5.1.15
+
 ## [10.1.9](https://github.com/libp2p/js-libp2p/compare/tcp-v10.1.8...tcp-v10.1.9) (2025-04-09)